fix unique temp files and deletion

+++ __tests__/read-toml.test.ts
@@ -5,41 +5,57 @@
 
 import { processTOMLFile } from '../src/read-toml'
 import { expect } from '@jest/globals'
-
+import { v4 as uuidv4 } from 'uuid';
 import * as fs from 'fs'
 import * as path from 'path'
 
 describe('TOML Reader', () => {
-  const tempDir = path.join(__dirname, 'temp')
-  const tempFilePath = path.join(tempDir, 'file.toml')
-  const tomlContent = 'example = { value = "test" }'
+  const tempDir = process.env.RUNNER_TEMP || path.join(__dirname, 'temp');
+  let tempFilePath: string;
+  let uniqueFilename: string;
 
   beforeAll(() => {
-    if (!fs.existsSync(tempDir)) {
-      fs.mkdirSync(tempDir);
+    // If not within Github Runner, create the temp directory if it doesn't exist
+    if (!process.env.RUNNER_TEMP && !fs.existsSync(tempDir)) {
+        fs.mkdirSync(tempDir);
     }
-    fs.writeFileSync(tempFilePath, tomlContent);
   });
 
   afterAll(() => {
-    if (fs.existsSync(tempFilePath)) {
-      fs.unlinkSync(tempFilePath);
-    }
-    fs.rm(tempDir, { recursive: true }, err => {
-      if (err) {
-        console.error(err);
+    if (!process.env.RUNNER_TEMP) {
+      try {
+        fs.rmSync(tempDir, { recursive: true });
+      } catch (error) {
+        console.error(`Error removing temporary directory: ${error}`);
       }
-    });
+    }
+  });
+
+  beforeEach(() => {
+    uniqueFilename = `file-${uuidv4()}.toml`;
+    tempFilePath = path.join(tempDir, uniqueFilename);
+  });
+
+  afterEach(() => {
+    if (!process.env.RUNNER_TEMP) {
+      try {
+        if (fs.existsSync(tempFilePath)) {
+          fs.unlinkSync(tempFilePath);
+        }
+      } catch (error) {
+        console.error(`Error removing temporary file: ${error}`);
+      }
+    }
   });
 
   it('should read a field from a TOML file', async () => {
-    const fieldPath = 'example.value'
-    const expectedValue = 'test'
-    fs.writeFileSync(tempFilePath, tomlContent)
+    const fieldPath = 'example.value';
+    const expectedValue = 'test';
+    fs.writeFileSync(tempFilePath, 'example = { value = "test" }');
 
-    const value = await processTOMLFile(tempFilePath, fieldPath)
-    expect(value).toEqual(expectedValue)
-  })
+    const value = await processTOMLFile(tempFilePath, fieldPath);
+    expect(value).toEqual(expectedValue);
+  });
 
   it('should throw an error if the file does not exist', async () => {
     const filePath = 'nonexistent.toml'
@@ -51,70 +67,52 @@ describe('TOML Reader', () => {
   })
 
   it('should throw an error for malformed TOML content', async () => {
-    const dirPath = 'path/to'
-    const filePath = `${dirPath}/malformed.toml`
     const fieldPath = 'example.value'
+    //? const expectedValue = 'test'
+    fs.writeFileSync(tempFilePath, 'invalid = :malformed_content')
 
-    fs.mkdirSync(dirPath, { recursive: true })
-    fs.writeFileSync(filePath, 'invalid = :malformed_content')
-
-    await expect(processTOMLFile(filePath, fieldPath)).rejects.toThrow(
+    await expect(processTOMLFile(tempFilePath, fieldPath)).rejects.toThrow(
       /Unexpected character, expecting string, number, datetime, boolean, inline array or inline table/
     )
 
-    // Clean up
-    fs.unlinkSync(filePath)
   })
 
   it.skip('should throw an error if the field is not found in the TOML file', async () => {
-    const dirPath = 'path/to'
-    const filePath = `${dirPath}/valid.toml`
     const fieldPath = 'nonexistent.field'
 
-    fs.mkdirSync(dirPath, { recursive: true })
-    fs.writeFileSync(filePath, 'example = { value = "test" }')
+    fs.writeFileSync(tempFilePath, 'example = { value = "test" }')
 
-    await expect(processTOMLFile(filePath, fieldPath)).rejects.toThrow(
+    await expect(processTOMLFile(tempFilePath, fieldPath)).rejects.toThrow(
       "Field 'nonexistent.field' not found in the TOML file"
     )
 
-    // Clean up
-    fs.unlinkSync(filePath)
-    fs.rm(dirPath, { recursive: true }, err => {
-      if (err) {
-        console.error(err)
-      }
-    })
   })
 
   it.skip('should correctly retrieve a nested field from the TOML file', async () => {
-    const filePath = 'path/to/nested.toml'
     const fieldPath = 'nested.example.value'
     const expectedValue = 'nestedTest'
-    fs.writeFileSync(filePath, '[nested.example]\nvalue = "nestedTest"')
 
-    const value = await processTOMLFile(filePath, fieldPath)
+    fs.writeFileSync(tempFilePath, '[nested.example]\nvalue = "nestedTest"')
+
+    const value = await processTOMLFile(tempFilePath, fieldPath)
+  
     expect(value).toEqual(expectedValue)
-    fs.unlinkSync(filePath)
+  
   })
 
   it.skip('should handle non-string field values correctly', async () => {
-    const filePath = 'path/to/variety.toml'
     const fieldPathNumber = 'example.number'
     const fieldPathBoolean = 'example.boolean'
-    fs.writeFileSync(filePath, 'example = { number = 42, boolean = true }')
+    fs.writeFileSync(tempFilePath, 'example = { number = 42, boolean = true }')
 
-    const numberValue = await processTOMLFile(filePath, fieldPathNumber)
+    const numberValue = await processTOMLFile(tempFilePath, fieldPathNumber)
     expect(numberValue).toEqual('42') // Assuming conversion to string
 
-    const booleanValue = await processTOMLFile(filePath, fieldPathBoolean)
+    const booleanValue = await processTOMLFile(tempFilePath, fieldPathBoolean)
     expect(booleanValue).toEqual('true') // Assuming conversion to string
-
-    fs.unlinkSync(filePath)
   })
 
   it.skip('throws an error if the field value is not a string', async () => {
-    const filePath = '/path/to/valid.toml'
     const fieldPath = 'example.nonString'
 
     jest.spyOn(fs, 'existsSync').mockReturnValue(true)
@@ -122,8 +120,9 @@ describe('TOML Reader', () => {
       .spyOn(fs, 'readFileSync')
       .mockReturnValue('example = { nonString = 123 }')
 
-    await expect(processTOMLFile(filePath, fieldPath)).rejects.toThrow(
+    await expect(processTOMLFile(tempFilePath, fieldPath)).rejects.toThrow(
       `Expected a string at '${fieldPath}', but found a different type`
     )
+
   })
 })
